Fix bug #90d85b
This patch adds a feature that causes the timer state to be updated
whenever a window focus change occurs. In the new scheme, the timer
is only running whenever a tab that is playing audio or is the active
tab in a focused window matches a rule in the match list.