Prevent race condition in timer update
Sometimes, when a new entry is added to `matches' that fits one of
the Timer Activation Conditions, the on-page timer will check the
state of the timer before it has been updated, so the timer will
start, but the on-page timer will not.

This patch updates the implementation to use Chrome message passing
to update the time element asynchronously and eliminate the race
condition.